+++ cmd/obitools/obiuniq/main.go
@@ -30,6 +30,8 @@ func main() {
 	// trace.Start(ftrace)
 	// defer trace.Stop()
 
+	obioptions.SetBatchSize(10)
+	obioptions.SetReadQualities(false)
 	optionParser := obioptions.GenerateOptionParser(obiuniq.OptionSet)
 
 	_, args := optionParser(os.Args)