+++ pkg/obioptions/options.go
@@ -30,6 +30,7 @@ var _PprofMudex = 10
 var _PprofGoroutine = 6060
 var _Quality_Shift_Input = byte(33)
 var _Quality_Shift_Output = byte(33)
+var _Read_Qualities = true
 
 var __taxdump__ = ""
 var __alternative_name__ = false
@@ -291,6 +292,10 @@ func CLIBatchSize() int {
 	return _BatchSize
 }
 
+func CLIReadQualities() bool {
+	return _Read_Qualities
+}
+
 // SetDebugOn sets the debug mode on.
 func SetDebugOn() {
 	_Debug = true
@@ -299,6 +304,10 @@ func SetDebugOn() {
 // SetDebugOff sets the debug mode off.
 func SetDebugOff() {
 	_Debug = false
+}
+
+func SetReadQualities(status bool) {
+	_Read_Qualities = status
 }
 
 // SetWorkerPerCore sets the number of workers per CPU core.