+++ pkg/obiseq/biosequence.go
@@ -12,6 +12,7 @@ package obiseq
 
 import (
 	"crypto/md5"
+	"encoding/hex"
 	"slices"
 	"sync"
 	"sync/atomic"
@@ -182,6 +183,9 @@ func (s *BioSequence) Copy() *BioSequence {
 	newSeq.sequence = CopySlice(s.sequence)
 	newSeq.qualities = CopySlice(s.qualities)
 	newSeq.feature = CopySlice(s.feature)
+	if s.revcomp != nil {
+		newSeq.revcomp = s.revcomp.Copy()
+	}
 
 	if len(s.annotations) > 0 {
 		s.annot_lock.Lock()
@@ -373,6 +377,11 @@ func (s *BioSequence) Source() string {
 // Returns [16]byte, the MD5 hash of the BioSequence.
 func (s *BioSequence) MD5() [16]byte {
 	return md5.Sum(s.sequence)
+}
+
+func (s *BioSequence) MD5String() string {
+	md5_hash := s.MD5()
+	return hex.EncodeToString(md5_hash[:])
 }
 
 // SetId sets the id of the BioSequence.